make Dynamic a built-in type
