+++ std_lib/sqlite.ard
@@ -1,17 +1,17 @@
 use ard/decode
 
-private extern fn _create_connection(filepath: Str) decode::Dynamic!Str = "SqliteCreateConnection"
-private extern fn _close(conn: decode::Dynamic) Void!Str = "SqliteClose"
-private extern fn _exec(conn: decode::Dynamic, sql: Str) Void!Str = "SqliteExec"
-private extern fn _query(conn: decode::Dynamic, sql: Str) decode::Dynamic!Str = "SqliteQuery"
-private extern fn _first(conn: decode::Dynamic, sql: Str) decode::Dynamic!Str = "SqliteFirst"
-private extern fn _insert(conn: decode::Dynamic, table: Str, values: [Str:decode::Dynamic]) decode::Dynamic!Str = "SqliteInsert"
-private extern fn _update(conn: decode::Dynamic, table: Str, where: Str, values: [Str:decode::Dynamic]) decode::Dynamic!Str = "SqliteUpdate"
-private extern fn _delete(conn: decode::Dynamic, table: Str, where: Str) Bool!Str = "SqliteDelete"
-private extern fn _count(conn: decode::Dynamic, table: Str, where: Str) Int!Str = "SqliteCount"
+private extern fn _create_connection(filepath: Str) Dynamic!Str = "SqliteCreateConnection"
+private extern fn _close(conn: Dynamic) Void!Str = "SqliteClose"
+private extern fn _exec(conn: Dynamic, sql: Str) Void!Str = "SqliteExec"
+private extern fn _query(conn: Dynamic, sql: Str) Dynamic!Str = "SqliteQuery"
+private extern fn _first(conn: Dynamic, sql: Str) Dynamic!Str = "SqliteFirst"
+private extern fn _insert(conn: Dynamic, table: Str, values: [Str:Dynamic]) Dynamic!Str = "SqliteInsert"
+private extern fn _update(conn: Dynamic, table: Str, where: Str, values: [Str:Dynamic]) Dynamic!Str = "SqliteUpdate"
+private extern fn _delete(conn: Dynamic, table: Str, where: Str) Bool!Str = "SqliteDelete"
+private extern fn _count(conn: Dynamic, table: Str, where: Str) Int!Str = "SqliteCount"
 
 struct Database {
-  conn: decode::Dynamic
+  conn: Dynamic
   filepath: Str
 }
 
@@ -29,19 +29,19 @@ impl Database {
     _exec(@conn, sql)
   }
 
-  fn query(sql: Str) decode::Dynamic!Str {
+  fn query(sql: Str) Dynamic!Str {
     _query(@conn, sql)
   }
 
-  fn first(sql: Str) decode::Dynamic!Str {
+  fn first(sql: Str) Dynamic!Str {
     _first(@conn, sql)
   }
 
-  fn insert(table: Str, values: [Str:decode::Dynamic]) decode::Dynamic!Str {
+  fn insert(table: Str, values: [Str:Dynamic]) Dynamic!Str {
     _insert(@conn, table, values)
   }
 
-  fn update(table: Str, where: Str, values: [Str:decode::Dynamic]) decode::Dynamic!Str {
+  fn update(table: Str, where: Str, values: [Str:Dynamic]) Dynamic!Str {
     _update(@conn, table, where, values)
   }
 
@@ -58,7 +58,7 @@ impl Database {
     Result::ok(count > 0)
   }
 
-  fn upsert(table: Str, where: Str, values: [Str:decode::Dynamic]) decode::Dynamic!Str {
+  fn upsert(table: Str, where: Str, values: [Str:Dynamic]) Dynamic!Str {
     let found = try @exists(table, where)
     match found {
       true => @update(table, where, values),
